feat(model): support deepseek v3.2 (#316)

+++ src/parallax/metal/paged_attention/reshape_and_cache.metal
@@ -22,7 +22,7 @@ if (kv_head_dim_idx >= n_kv_heads * max_dim)
 int head_idx = kv_head_dim_idx / max_dim;
 int dim_idx = kv_head_dim_idx % max_dim;
 
-int64_t slot = slot_mapping[token_idx];
+int slot = slot_mapping[token_idx];
 
 // Handle padding tokens (slot == -1)
 if (slot < 0) {
@@ -30,8 +30,8 @@ if (slot < 0) {
 }
 
 int b_size = block_size;
-int64_t block_idx = slot / b_size;
-int64_t block_offset = slot % b_size;
+int block_idx = slot / b_size;
+int block_offset = slot % b_size;
 
 int l_idx = layer_idx;
 int n_blocks = num_blocks;
@@ -48,7 +48,7 @@ if (dim_idx < k_dim) {
     int64_t block_stride = n_kv_heads * head_stride;
     int64_t layer_stride = n_blocks * block_stride;
 
-    int64_t dest_idx = (int64_t)l_idx * layer_stride + block_idx * block_stride +
+    int64_t dest_idx = (int64_t)l_idx * layer_stride + (int64_t)block_idx * block_stride +
                        (int64_t)head_idx * head_stride + block_offset * k_dim +
                        dim_idx;
 
@@ -67,7 +67,7 @@ if (dim_idx < v_dim) {
     int64_t block_stride = n_kv_heads * head_stride;
     int64_t layer_stride = n_blocks * block_stride;
 
-    int64_t dest_idx = (int64_t)l_idx * layer_stride + block_idx * block_stride +
+    int64_t dest_idx = (int64_t)l_idx * layer_stride + (int64_t)block_idx * block_stride +
                        (int64_t)head_idx * head_stride + block_offset * v_dim +
                        dim_idx;
